Fixed errors moved all headers into one file
- suppressed warning for comparison between signed and unsigned ints
The vector takes integer values but .size() returns an unsigned integer. Type casting may mean that information could be lost on incorrect hence suppressed the warning.
- using size_t in place of int would not have worked since it isn't explicitly and integer.
