+++ src/headers/AirCraft_Carrier.h
@@ -8,10 +8,11 @@
 #ifndef AIRCRAFT_CARRIER_H
 #define AIRCRAFT_CARRIER_H
 
+#include "Ship_BaseClass.h"
 #include "Constants.h"
-#include "ships/Ship_BaseClass.h"
 
 class AirCraft_Carrier : public Ship_BaseClass {
+public:
 	AirCraft_Carrier() : Ship_BaseClass() {
 		setShipHealth(Constants::AIRCRAFT_CARRIER_SIZE);
 	}
