+++ src/headers/Submarine.h
@@ -1,5 +1,5 @@
 /*
- * Frigate.h
+ * Submarine.h
  *
  *  Created on: 22 Feb 2024
  *      Author: Charlie
@@ -8,8 +8,8 @@
 #ifndef SUBMARINE_H
 #define SUBMARINE_H
 
+#include "Ship_BaseClass.h"
 #include "Constants.h"
-#include "ships/Ship_BaseClass.h"
 
 class Submarine : public Ship_BaseClass {
 public:
